Update linting CICD (#10)
This pull request updates the checkout code step with the latest code
changes. Version 2 is no longer supported, so we have made the necessary
changes to ensure that the code works as expected with the current
version.

We have also added a requirements.txt file to the main branch, which
includes the pylint library. This will simplify the installation process
for new contributors and help prevent future bugs in the code. By
including all required libraries in the requirements.txt file, we can
ensure that all necessary libraries are installed at once. This will
save time and reduce the risk of errors.