Entry group name (#34)
In this PR there is a correction to the entry group id and tag template.

The entry group ID is now defined using a regular expression. It
restricts the input to letters, numbers, and underscores, allowing a
maximum of 119 characters for each table and 120 for each dataset for
bigquery (as the maximum is 225 char for the tag template id) and 50
characters to cloudsql insatnce_id (as the maximum of the entry group id
is 65). This update ensures that the argument for the entry group ID and
the tag template id is generic and can accommodate any ID and both use
cases, giving the user as much information as possible in the entry
group id.