Parallelized Table Inspection and Block Processing with Apache Beam (#44)
This pull request introduces parallelization to the code, enabling the
inspection of multiple tables and parallel processing of blocks within
each table. The implementation utilizes Apache Beam, allowing for
efficient and scalable inspection of tables. Additionally, a Dataflow
template is generated, facilitating the execution of the pipeline on
Google Cloud Dataflow.

---------

Co-authored-by: Carla <carlablannn@gmail.com>